+++ Lab 2/src/Complex.h
@@ -14,9 +14,9 @@ using namespace std;
 class Complex {
 	public:
 		Complex();
-		Complex(int r, int i);
-		int getimaginary();
-		int getreal();
+		Complex(float r, float i);
+		float getimaginary();
+		float getreal();
 		friend Complex operator +(const Complex& coma, const Complex& comb);
 		friend ostream& operator <<(ostream& ost, const Complex &c);
 		void operator =(Complex c);
